requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
webvtt-py>=0.4.6
assemblyai>=0.43.0
ffmpeg-python>=0.2.0
requests>=2.31.0
httpx>=0.25.0
reportlab>=4.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0